bq_conversational_agent/requirements.txt
google-adk
google-cloud-bigquery
google-cloud-aiplatform[adk,agent_engines]
functions-framework
cloudpickle
uvicorn
fastapi
google-cloud-geminidataanalytics
